+++ src/DrawioAppServer/DrawioAppServer.ts
@@ -1,0 +1,6 @@
+import { Webview } from "vscode";
+import { DrawioInstance } from "../DrawioInstance";
+
+export interface DrawioAppServer {
+	setupWebview(webview: Webview): Promise<DrawioInstance>;
+}