+++ src/DrawioAppServer/ExternallyHostedDrawioAppServer.ts
@@ -1,0 +1,11 @@
+import { HostedDrawioAppServer } from "./HostedDrawioAppServer";
+
+export class ExternallyHostedDrawioAppServer extends HostedDrawioAppServer {
+	constructor(public url: string) {
+		super();
+	}
+
+	public async getIndexUrl(): Promise<string> {
+		return this.url;
+	}
+}