+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 ## 0.2.0
 
+-   Implements offline mode (enabled by default).
+-   Implements config to disable offline mode.
+-   Implements config to chose a custom drawio url.
+
+## 0.1.3
+
 -   Treats \*.dio files the same as \*.drawio files.
+-   Makes extension compatible with VS Code 1.44.
 
 ## 0.1.0